% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_cycle_time.R
\name{estimate_cycle_time}
\alias{estimate_cycle_time}
\title{Estimate cycle time}
\usage{
estimate_cycle_time(
  exprs,
  ensembl_ids = NULL,
  entrez_ids = NULL,
  handle_multiple_observations = c("mean", "median", "max", "remove"),
  return_full_residuals = FALSE,
  model = c("full", "secretory"),
  quiet = FALSE
)
}
\arguments{
\item{exprs}{Expression matrix with rows as gene observations and columns as samples}

\item{ensembl_ids}{Optional vector of Ensembl gene identifiers corresponding to each row of the expression matrix \code{exprs}.}

\item{entrez_ids}{Optional vector of Entrez gene identifiers corresponding to each row of the expression matrix \code{exprs}.}

\item{handle_multiple_observations}{Method of handling multiple observations correspond to the same gene. \code{"remove"} will remove all observations of multiple observations, while \code{"mean"}, \code{"median"}, and \code{"max"} will consolidate multiple observations into one using the specified function.}

\item{return_full_residuals}{Return a list of residual matrices for all samples if \code{TRUE}. The matrices contain residual values for every time point for all genes.}

\item{model}{Use full model (recommended) or secretory model.}

\item{quiet}{Run quietly and don't output messages.}
}
\value{
Returns a list with the following items:
\describe{
  \item{estimated_time}{A vector of estimated cycle times from 0 to 99 for each sample.}
  \item{mse}{A matrix of mean squared error values for each sample at each timepoint.}
  \item{residuals}{A matrix of residual values after subtracting the cycle stage effect.}
  \item{full_residuals}{If \code{return_full_residuals=TRUE}, a list of matrices containing residual values for all timepoints.}
}
}
\description{
Estimate uterine cycle time of endometrium samples using expression data.
}
\examples{
data(example_rna_exprs)
results <- estimate_cycle_time(exprs=example_rna_exprs)
results$estimated_time

data(example_array_exprs)
data(example_array_entrez_ids)
results <- estimate_cycle_time(exprs=example_array_exprs, entrez_ids=example_array_entrez_ids)
results$estimated_time
}
