% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_array_entrez_ids}
\alias{example_array_entrez_ids}
\title{Example Entrez IDs}
\format{
A vector of length 5000 containing Entrez IDs.
}
\usage{
example_array_entrez_ids
}
\description{
List of probe IDs to Entrez IDs corresponding to the 5000 probes in the example_array_exprs dataset.
}
\keyword{datasets}
