% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_array_exprs}
\alias{example_array_exprs}
\title{Example array data}
\format{
A matrix with 3 columns and 5000 rows.
}
\usage{
example_array_exprs
}
\description{
Example Illumina microarray expression data of 3 samples and 5000 probes.
}
\keyword{datasets}
