% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_rna_exprs}
\alias{example_rna_exprs}
\title{Example RNA-seq data}
\format{
A matrix with 3 columns and 2000 rows.
}
\usage{
example_rna_exprs
}
\description{
Example RNA-seq count data of 3 samples and 2000 genes.
}
\keyword{datasets}
