% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mse.R
\name{plot_mse}
\alias{plot_mse}
\title{Plot MSE curve}
\usage{
plot_mse(x, sample_name, ...)
}
\arguments{
\item{x}{Cycle time estimation list generated from the \code{estimate_cycle_time} function.}

\item{sample_name}{Sample name of sample to plot corresponding to one of the column names in the original expression matrix.}

\item{...}{Other optional arguments to be passed to \code{plot}.}
}
\value{
None
}
\description{
Plot the MSE curve of a sample using base graphics.
}
\examples{
# Simulate gene expression data for 100 genes and 3 samples
entrez_ids <- 9801:9900
y <- matrix(rnorm(100*3, 5, 2), nrow=100)
colnames(y) <- paste0("sample_", 1:3)
results <- estimate_cycle_time(exprs=y, entrez_ids=entrez_ids)
plot_mse(results, sample_name="sample_1")
}
